console_conf: fix showing error messages, when snap create-user fails
(cherry picked from commit 6de89705600cb86ae79be902da30abf434e2c759)
